fixed CG and add new file

+++ azheganova.faina/S1/convertfrominfixtopostfix.cpp
@@ -1,21 +1,7 @@
 #include "convertfrominfixtopostfix.h"
 #include <cctype>
 #include <iostream>
-#include "stack.h"
-
-
-bool isDigit(const std::string & string)
-{
-  try
-  {
-    std::stoll(string, nullptr, 10);
-  }
-  catch(const std::exception & e)
-  {
-    return false;
-  }
-  return true;
-}
+#include "checkingfordigit.h"
 
 bool isOperator(std::string oper)
 {
@@ -26,54 +12,60 @@ void convertFromInfixToPostfix(Queue< std::string > & queue, Stack< std::string 
 {
   while (!queue.isEmpty() || !stack.isEmpty())
   {
-    std::string element = queue.drop();
-    queue.pop();
-    if (isDigit(element))
+    if (!queue.isEmpty())
     {
-      postfix.push(element);
-    }
-    else if (isOperator(element))
-    {
-      if (element == "+" || element == "-")
+      std::string element = queue.drop();
+      queue.pop();
+      if (isDigit(element))
       {
-        while (!stack.isEmpty() && isOperator(stack.drop()))
+        postfix.push(element);
+      }
+      else if (isOperator(element))
+      {
+        if (element == "+" || element == "-")
+        {
+          while (!stack.isEmpty() && isOperator(stack.drop()))
+          {
+            postfix.push(stack.drop());
+            stack.pop();
+          }
+          stack.push(element);
+        }
+        else if (element == "*" || element == "/" || element == "%")
+        {
+          while (!stack.isEmpty() && ((stack.drop() == "*" || stack.drop() == "/" || stack.drop() == "%")))
+          {
+            postfix.push(stack.drop());
+            stack.pop();
+          }
+          stack.push(element);
+        }
+      }
+      else if (element == "(")
+      {
+        stack.push(element);
+      }
+      else if (element == ")")
+      {
+        while ((stack.drop() != "(") && (!stack.isEmpty()))
         {
           postfix.push(stack.drop());
           stack.pop();
         }
-        stack.push(element);
+        stack.pop();
       }
-      else if (element == "*" || element == "/" || element == "%")
+      else
       {
-        while (!stack.isEmpty() && ((stack.drop() == "*" || stack.drop() == "/" || stack.drop() == "%")))
-        {
-          postfix.push(stack.drop());
-          stack.pop();
-        }
-        stack.push(element);
+        throw std::logic_error("error");
       }
     }
-    else if (element == "(")
+    else
     {
-      stack.push(element);
-    }
-    else if (element == ")")
-    {
-      while ((stack.drop() != "(") && (!stack.isEmpty()))
+      while (!stack.isEmpty())
       {
         postfix.push(stack.drop());
         stack.pop();
       }
-      stack.pop();
     }
-    else
-    {
-      throw std::logic_error("error");
-    }
-  }
-  while (!stack.isEmpty())
-  {
-    postfix.push(stack.drop());
-    stack.pop();
   }
 }
